fix(pois): explicit float8 casts in ST_MakePoint to resolve type inference error
PostgreSQL could not deduce a consistent type for $3/$4 when the same
parameters were used both as column values and inside ST_MakePoint().
Adding ::float8 casts resolves the "inconsistent types deduced" error
that appeared on first successful POI insert.

+++ backend/src/services/googlePoiService.ts
@@ -177,7 +177,7 @@ async function upsertPois(places: any[], poiType: string, dept: string, client: 
     await client.query(
       `INSERT INTO poi_cache (name, poi_type, lat, lng, geometry, tags, source, fetched_at)
        VALUES ($1, $2, $3, $4,
-               ST_SetSRID(ST_MakePoint($4, $3), 4326),
+               ST_SetSRID(ST_MakePoint($4::float8, $3::float8), 4326),
                $5, 'google_places', NOW())`,
       [
         (place.name ?? '').slice(0, 255),
